pyproject.toml
[project]
name = "wolf-creek-pass"
version = "0.2.0"
description = "Route-aware road conditions dashboard for Wolf Creek Pass (SR-35), powered by Claude Vision"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "boto3>=1.35.0",
    "click>=8.1.0",
    "googlemaps>=4.10.0",
    "polyline>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "schedule>=1.2.0",
]

[project.scripts]
wolf-creek = "traffic_cam_monitor:cli"
wolf-creek-query = "query_data:cli"

[tool.poe.tasks]
monitor = { cmd = "python traffic_cam_monitor.py", help = "Start the traffic camera monitor (runs hourly)" }
"monitor:once" = { cmd = "python traffic_cam_monitor.py --once", help = "Run a single capture cycle" }
query = { cmd = "python query_data.py recent", help = "Show recent captures" }
query-snow = { cmd = "python query_data.py snow", help = "Show snow conditions" }
query-animals = { cmd = "python query_data.py animals", help = "Show animal sightings" }
query-cycles = { cmd = "python query_data.py cycles", help = "Show capture cycle history" }
query-route = { cmd = "python query_data.py route", help = "Show current route info" }
dev = { cmd = "npm run dev", cwd = "frontend", help = "Start Vue dev server (Vite)" }
build = { cmd = "npm run build", cwd = "frontend", help = "Build Vue for production" }
localstack-up = { cmd = "docker compose up -d", help = "Start LocalStack" }
localstack-down = { cmd = "docker compose down", help = "Stop LocalStack" }
cdk-deploy = { shell = "cd infra && npx aws-cdk deploy --require-approval never", help = "Deploy AWS infrastructure" }
cdk-diff = { shell = "cd infra && npx aws-cdk diff", help = "Preview infrastructure changes" }

[dependency-groups]
dev = [
    "poethepoet>=0.31.0",
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
]
